feat(Install-PSModules): add Pscx module and modify installation parameters
This commit introduces two changes to the `Install-PSModules` function in the `Install-ExternalDependencies.ps1` script.

Firstly, it adds 'Pscx' to the list of modules to be installed. The Pscx (PowerShell Community Extensions) module provides a wide range of additional cmdlets, providers, aliases, filters, functions, and scripts that extend the capabilities of Windows PowerShell.

Secondly, it modifies the `Install-Module` command by adding `-AllowClobber` and `-SkipPublisherCheck` parameters. The `-AllowClobber` parameter allows you to install a module that has cmdlets with the same names as cmdlets in another module already installed on your system. The `-SkipPublisherCheck` parameter bypasses the publisher check when installing the module. These changes provide more flexibility and efficiency in the module installation process.