chore: clean up code formatting and refactor functions for modularity
- Remove unnecessary whitespace in `Import-FromInMemoryModule.ps1`.
- Rename `Unprotect-ExcelSheet` to `Unprotect-Excel` for clarity.
- Refactor `Unprotect-Excel` to dynamically import `Write-Logg` function from a remote script.
- Update `Write-Logg` to include dynamic importing of dependency installation logic.

This commit cleans up the code by removing extraneous whitespace and improves the modularity of the PowerShell scripts. The `Unprotect-Excel` function has been renamed for better clarity and now imports the `Write-Logg` function dynamically from a remote URI. Additionally, the `Write-Logg` function itself has been updated to dynamically import a new `Install-Dependencies` function, which handles the installation of required modules like 'pansies'. This change reduces the need for manual updates and maintenance of the dependency installation process within the script.