feat(logging): optimize dynamic logging module loading and improve backup handling
This commit introduces several enhancements to the Unprotect-Excel function. It optimizes the loading of the dynamic logging module by checking if it's already loaded before importing, and defines a unique name for the module to prevent conflicts. The backup process now compresses the original Excel file into a ZIP archive for space efficiency. Error handling is improved when extracting the Excel file with more descriptive messages, including the use of a new Test-MAMEncryption function to detect MDM encryption and provide entropy information. Additionally, the re-creation of the Excel file now uses native PowerShell cmdlets instead of .NET classes for better compatibility and simplicity.