refactor: Update Get-FileDownload.ps1 to allow custom Aria2c executable path
This commit modifies the Get-FileDownload.ps1 file to allow specifying a custom Aria2c executable path. Previously, the script would automatically search for the Aria2c executable in the 'c:\aria2c\' directory. Now, the script accepts the path as a parameter, giving more flexibility to the user. This change improves the usability of the script and allows users to use a different Aria2c executable if desired.