refactor: Update PowerShell profile script to use latest version of Aria2c
This commit updates the PowerShell profile script to use the latest version of Aria2c. Previously, the script would search for the Aria2c executable in the 'c:\aria2c\' directory. Now, it will download the latest version from the community repository and use that instead. This change improves the script's functionality by ensuring that users have access to the most up-to-date version of Aria2c.