refactor: Update Write-Logg.ps1 to improve console output
This commit modifies the Write-Logg.ps1 script to improve the console output. It adds a condition to check if the log level is 'VERBOSE' and outputs the log message to the console in green color. This change enhances the functionality and usability of the script by providing better visibility of verbose log messages.