refactor: Update Get-KapeAndTools.ps1 to dynamically set KAPE install location
This commit modifies the Get-KapeAndTools.ps1 script to dynamically set the KAPE install location. It checks if the specified location exists using the Resolve-Path cmdlet, and if not, it falls back to a wildcard search for the install location. This change improves the functionality and usability of the script by ensuring that the correct KAPE install location is used.