refactor: Add Get-FFmpeg function to download and install FFmpeg
This commit adds the Get-FFmpeg function to the repository. The function is used to download and install FFmpeg from a specified URL. It also provides the option to download FFmpeg locally and extract it. This new function enhances the functionality of the repository by providing an easy way to install FFmpeg.